requirements.txt
# Core
numpy>=1.23
pandas>=1.5
tqdm>=4.65
joblib>=1.3

# Machine Learning
scikit-learn>=1.3

# Deep Learning
torch>=2.0
torchvision>=0.15

# NLP - Transformers
transformers>=4.36
sentencepiece>=0.1.99

# Computer Vision
opencv-python>=4.8
Pillow>=9.5

# OCR
pytesseract>=0.3.10

# Web App
streamlit>=1.28
